fix: cast usize to u32 for tree-sitter 0.26 named_child API
tree-sitter 0.26.11 changed named_child() to take u32 instead of
usize. Cast loop indices with 'i as u32' at all call sites.

+++ crates/colab-lang-go/src/imports.rs
@@ -213,7 +213,7 @@ pub fn ensure(target: &str, source_code: &str) -> String {
     let root = tree.root_node();
     let mut insert_at: Option<usize> = None;
     for i in 0..root.named_child_count() {
-        if let Some(child) = root.named_child(i)
+        if let Some(child) = root.named_child(i as u32)
             && child.kind() == "package_clause"
         {
             let end = child.end_byte();
