+++ crates/colab-lang-rust/src/uses.rs
@@ -232,7 +232,7 @@ pub fn ensure(target: &str, source_code: &str) -> String {
     let root = tree.root_node();
     let mut insert_at = 0usize;
     for i in 0..root.named_child_count() {
-        let Some(child) = root.named_child(i) else {
+        let Some(child) = root.named_child(i as u32) else {
             break;
         };
         if child.kind() == "inner_attribute_item" {
